+++ source/js/common/fhir-batch-query.js
@@ -1,4 +1,5 @@
 import { updateUrlWithParam } from './utils';
+import definitionsIndex from '../search-parameters/definitions';
 
 let commonRequestCache = {}; // Map from url to result JSON
 
@@ -30,6 +31,7 @@ export class FhirBatchQuery {
     this._maxPerBatch = maxRequestsPerBatch;
     this._maxActiveReq = maxActiveRequests;
     this._activeReq = [];
+    this._onChangeServiceBaseUrlListeners = [];
     // Timeout between requests in milliseconds
     // (=0, if there is no timeout between requests):
     this._msBetweenRequests = 0;
@@ -37,10 +39,168 @@ export class FhirBatchQuery {
     this._lastSuccessTime = Date.now();
     // The client side should give up if no successful response in 90 seconds
     this._giveUpTimeout = 90 * 1000;
+    // NCBI E-utilities API Key
+    this._apiKey = '';
   }
 
   getServiceBaseUrl() {
     return this._serviceBaseUrl;
+  }
+
+  /**
+   * @typedef FHIRServerFeatures
+   * @type {Object}
+   * @property {boolean} [sortObservationsByDate] - whether sorting
+   *           Observations by date available
+   * @property {boolean} [sortObservationsByAgeAtEvent] - whether sorting
+   *           Observations by age-at-event available
+   */
+
+  /**
+   * Returns an object describing the server features.
+   * @return {FHIRServerFeatures}
+   */
+  getFeatures() {
+    return this._features || {};
+  }
+
+  /**
+   * Return version name e.g. "R4"
+   * @return {string}
+   */
+  getVersionName() {
+    return this._versionName;
+  }
+
+  /**
+   * Initialize/reinitialize FhirBatchQuery instance.
+   * @param {string} [newServiceBaseUrl] - new FHIR REST API Service Base URL
+   *                 (https://www.hl7.org/fhir/http.html#root)
+   * @return {Promise}
+   */
+  initialize(newServiceBaseUrl) {
+    if (newServiceBaseUrl && newServiceBaseUrl !== this._serviceBaseUrl) {
+      this.clearPendingRequests();
+      this._serviceBaseUrl = newServiceBaseUrl;
+      delete this._initializationPromise;
+    }
+
+    if (!this._initializationPromise) {
+      const initializationRequests = [
+        this.getWithCache('metadata', { combine: false }),
+        this.getWithCache('Observation?_sort=date&_elements=id&_count=1', {
+          combine: false
+        }),
+        this.getWithCache(
+          'Observation?_sort=age-at-event&_elements=id&_count=1',
+          { combine: false }
+        )
+      ];
+
+      this._initializationPromise = Promise.all(
+        // Convert reject to resolve to emulate Promise.allSettled behaviour (for Edge/IE11)
+        initializationRequests.map((promise) =>
+          promise.catch(() => {
+            return false;
+          })
+        )
+      ).then(
+        ([
+          metadata,
+          observationsSortedByDate,
+          observationsSortedByAgeAtEvent
+        ]) => {
+          const fhirVersion = metadata.data.fhirVersion;
+          this._versionName = getVersionNameByNumber(fhirVersion);
+          if (!this._versionName) {
+            return Promise.reject({
+              error: 'Unsupported FHIR version: ' + fhirVersion
+            });
+          }
+          this._features = {
+            sortObservationsByDate: observationsSortedByDate.data.total > 0,
+            sortObservationsByAgeAtEvent: !!observationsSortedByAgeAtEvent
+          };
+
+          this._onChangeServiceBaseUrlListeners.forEach((fn) => fn());
+        }
+      );
+    }
+
+    return this._initializationPromise;
+  }
+
+  /**
+   * Adds a listener for the FHIR REST API Service Base URL change event
+   * and returns a function to remove this listener
+   * @param {Function} handler
+   * @return {Function}
+   */
+  addChangeServiceBaseUrlEventListener(handler) {
+    this._onChangeServiceBaseUrlListeners.push(handler);
+
+    return () => {
+      const index = this._onChangeServiceBaseUrlListeners.indexOf(handler);
+      if (index > -1) {
+        this._onChangeServiceBaseUrlListeners.splice(index, 1);
+      }
+    };
+  }
+
+  /**
+   * Sets the maximum number of requests that can be combined
+   * (1 - turn off combined requests)
+   * @param {number} val
+   */
+  setMaxRequestsPerBatch(val) {
+    this._maxPerBatch = val;
+  }
+
+  /**
+   * Gets the maximum number of requests that can be combined
+   * @return {number}
+   */
+  getMaxRequestsPerBatch() {
+    return this._maxPerBatch;
+  }
+
+  /**
+   * Sets the maximum number of requests that can be executed simultaneously
+   * @param {number} val
+   */
+  setMaxActiveRequests(val) {
+    this._maxActiveReq = val;
+  }
+
+  /**
+   * Gets the maximum number of requests that can be executed simultaneously
+   * @return {number}
+   */
+  getMaxActiveRequests() {
+    return this._maxActiveReq;
+  }
+
+  /**
+   * Sets the NCBI E-utilities API Key.
+   * See https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/
+   * @param {string} val
+   */
+  setApiKey(val) {
+    this._apiKey = (val || '').trim();
+  }
+
+  /**
+   * Adds the an API Key to the passed URL and returns this new URL.
+   * @param {string} url
+   * @return {string}
+   */
+  addApiKeyToUrl(url) {
+    const apiKeyParam = 'api_key=' + encodeURIComponent(this._apiKey);
+    if (url.indexOf('?') === -1) {
+      return url + '?' + apiKeyParam;
+    } else {
+      return url + '&' + apiKeyParam;
+    }
   }
 
   static clearCache() {
@@ -50,13 +210,14 @@ export class FhirBatchQuery {
   /**
    * Gets the response content from a URL.
    * @param {string} url - the URL whose data is to be retrieved.
+   * @param {boolean} combine - whether to combine requests in a batch
    * @return {Promise} resolves/rejects with Object {status, data}, where
    *                   status is HTTP status number,
    *                   data is Object constructed from a JSON response
    */
-  get(url) {
+  get(url, { combine = true }) {
     return new Promise((resolve, reject) => {
-      this._pending.push({ url, resolve, reject });
+      this._pending.push({ url, combine, resolve, reject });
       if (this._pending.length < this._maxPerBatch) {
         clearTimeout(this._batchTimeoutId);
         this._batchTimeoutId = setTimeout(
@@ -129,12 +290,21 @@ export class FhirBatchQuery {
   /**
    * Sends XMLHttpRequest
    * @private
+   * @param {Object} settings - a set of key/value pairs that configure
+   *                 the XMLHttpRequest request
+   * @param {string} settings.method - HTTP method
+   * @param {string} settings.url - request URL
+   * @param {string} settings.body - request body if method === 'POST'
+   * @param {boolean} settings.combine - whether to combine requests in a batch
+   * @param {string} settings.contentType - Content-Type request header value
+   * @param {string} settings.logPrefix - prefix for console log messages
    * @return {Promise}
    */
   _request({
     method = 'GET',
     url,
-    body,
+    body = undefined,
+    combine = true,
     contentType = 'application/fhir+json',
     logPrefix = ''
   }) {
@@ -172,6 +342,7 @@ export class FhirBatchQuery {
               method,
               url,
               body,
+              combine,
               contentType,
               logPrefix,
               resolve,
@@ -200,6 +371,10 @@ export class FhirBatchQuery {
         }
       };
 
+      if (this._apiKey) {
+        url = this.addApiKeyToUrl(url);
+      }
+
       oReq.open(method, url);
       oReq.timeout = this._giveUpTimeout;
       oReq.setRequestHeader('Content-Type', contentType);
@@ -217,13 +392,17 @@ export class FhirBatchQuery {
     let requests = [];
 
     if (this._pending.length) {
-      if (this._pending[0].method === 'POST') {
+      if (
+        this._pending[0].method === 'POST' ||
+        this._pending[0].combine === false
+      ) {
         requests.push(this._pending.shift());
       } else {
         while (
           this._pending.length &&
           this._maxPerBatch > requests.length &&
-          this._pending[0].method !== 'POST'
+          this._pending[0].method !== 'POST' &&
+          this._pending[0].combine === true
         ) {
           requests.push(this._pending.shift());
         }
@@ -323,6 +502,7 @@ export class FhirBatchQuery {
       request.abort();
     });
     this._activeReq = [];
+    this._onChangeServiceBaseUrlListeners = [];
   }
 
   getFullUrl(url) {
@@ -349,11 +529,14 @@ export class FhirBatchQuery {
   /**
    * Like "get", but uses a cache if the URL has been requested before.
    * @param {string} url - the URL whose data is to be retrieved.
+   * @param {Object} options - additional options
+   * @param {boolean} options.combine - whether to combine requests in a batch,
+   *                  true by default
    * @return {Promise} resolves/rejects with Object {status, data}, where
    *                   status is HTTP status number,
    *                   data is Object constructed from a JSON response
    */
-  getWithCache(url) {
+  getWithCache(url, options = { combine: true }) {
     return new Promise((resolve, reject) => {
       const fullUrl = this.getFullUrl(url),
         cachedReq = commonRequestCache[fullUrl];
@@ -361,7 +544,7 @@ export class FhirBatchQuery {
         console.log('Using cached data');
         resolve(cachedReq);
       } else {
-        this.get(fullUrl).then((result) => {
+        this.get(fullUrl, options).then((result) => {
           commonRequestCache[fullUrl] = result;
           resolve(result);
         }, reject);
@@ -493,4 +676,28 @@ export class FhirBatchQuery {
       }, reject);
     });
   }
+}
+
+/**
+ * Returns version name by version number or null if version number is not supported.
+ * @example
+ * // calling a function as shown below will return this string: 'R4'
+ * getVersionNameByNumber('4.0.1')
+ * @param versionNumber
+ * @return {string|null}
+ */
+export function getVersionNameByNumber(versionNumber) {
+  let versionName = null;
+
+  Object.keys(definitionsIndex.versionNameByVersionNumberRegex).some(
+    (versionRegEx) => {
+      if (new RegExp(versionRegEx).test(versionNumber)) {
+        versionName =
+          definitionsIndex.versionNameByVersionNumberRegex[versionRegEx];
+        return true;
+      }
+    }
+  );
+
+  return versionName;
 }